funcs.py Most functions that calculate agreement and majority votes in funcs rely on objects inside the data structure that holds all of the classifiers. That design is definitely not good and I'm rewriting all of the functions so that they do not rely on that.
trainingUnit.py
This unit now checks whether models already exist or not to load them. Also initially I was storing all the data in one dictionary but that made the program ran out of battery. Because of that, instead now I'm reading the individual data set that I'm going to test at each time. Also it is important to note that I'm not using the random split but instead training on the entire data set available for each participant.